Added Best/Worst Categories to Progress Screen
Added a section to the Progress screen that details the units of emissions based on category. Section is labeled as 'Key Factors' and is intended to inform the user on the factors that contribute most to their carbon footprint. An alert icon will be displayed when the emission from a category is high. Future updates to this feature will allow each category to be expanded on to show suggestions to the user on how to lower emissions for said category, and general information on what sources of carbon emission contribute to said category.