Added Quiz Screen
Added quizScreen to forum tab. Currently this is the only screen in the stack, so its the main component. The quiz screen reads from the local api and has an error message and shows as loading if this is not active. Once it fetches the data it goes through all questions, incrementing score if correct, then displaying done when there are no more questions.