This is a new branch with a bunch of other commits, each one is described below
Need to add more filtering to only return certain things (The numbers needed to load in the log)

This took a lot of work since I had to mess around with asynchronous functions to make react wait properly.

Parsed the date as well as fixed my state hooks. Made the data returned from the server be of the right type.

Still need to parse the data and put it into arrays.

Parsed all the necessary data from the server. It does quite a bit but in summary we compare dates and return the right values for the 2d array based on that.

Added error check in log.js